Fix all type conversion warnings, plus misc. other stuff.

+++ Archiving/Animation Exporters/PXAnimatedGifExporter.h
@@ -22,7 +22,7 @@
 }
 
 - initWithSize:(NSSize)size iterations:(int)iterations;
-- (int)writeHeaderWithSize:(NSSize)size usingColorMap:(ColorMapObject *)colorMap ofSize:(int)numberOfColors withTransparentColor:(int)color;
+- (size_t)writeHeaderWithSize:(NSSize)size usingColorMap:(ColorMapObject *)colorMap ofSize:(int)numberOfColors withTransparentColor:(int)color;
 - (NSColor *)writeCanvas:(PXCanvas *)canvas withDuration:(NSTimeInterval)duration origin:(NSPoint)origin transparentColor:aColor;
 - (NSColor *)writeCanvas:(PXCanvas *)canvas withDuration:(NSTimeInterval)duration transparentColor:aColor;
 - (void)finalizeExport;
