+++ Other Sources/Foreign Code/CTGradient/CTGradient.m
@@ -18,16 +18,16 @@
 - (CTGradientElement *)elementAtIndex:(unsigned)index;
 
 - (CTGradientElement)removeElementAtIndex:(unsigned)index;
-- (CTGradientElement)removeElementAtPosition:(float)position;
+- (CTGradientElement)removeElementAtPosition:(CGFloat)position;
 @end
 
 //C Fuctions for color blending
-static void linearEvaluation   (void *info, const float *in, float *out);
-static void chromaticEvaluation(void *info, const float *in, float *out);
-static void inverseChromaticEvaluation(void *info, const float *in, float *out);
-static void transformRGB_HSV(float *components);
-static void transformHSV_RGB(float *components);
-static void resolveHSV(float *color1, float *color2);
+static void linearEvaluation   (void *info, const CGFloat *in, CGFloat *out);
+static void chromaticEvaluation(void *info, const CGFloat *in, CGFloat *out);
+static void inverseChromaticEvaluation(void *info, const CGFloat *in, CGFloat *out);
+static void transformRGB_HSV(CGFloat *components);
+static void transformHSV_RGB(CGFloat *components);
+static void resolveHSV(CGFloat *color1, CGFloat *color2);
 
 
 @implementation CTGradient
@@ -88,11 +88,11 @@ static void resolveHSV(float *color1, float *color2);
 	CTGradientElement *currentElement = elementList;
 	while(currentElement != nil)
 		{
-		[coder encodeValueOfObjCType:@encode(float) at:&(currentElement->red)];
-		[coder encodeValueOfObjCType:@encode(float) at:&(currentElement->green)];
-		[coder encodeValueOfObjCType:@encode(float) at:&(currentElement->blue)];
-		[coder encodeValueOfObjCType:@encode(float) at:&(currentElement->alpha)];
-		[coder encodeValueOfObjCType:@encode(float) at:&(currentElement->position)];
+		[coder encodeValueOfObjCType:@encode(CGFloat) at:&(currentElement->red)];
+		[coder encodeValueOfObjCType:@encode(CGFloat) at:&(currentElement->green)];
+		[coder encodeValueOfObjCType:@encode(CGFloat) at:&(currentElement->blue)];
+		[coder encodeValueOfObjCType:@encode(CGFloat) at:&(currentElement->alpha)];
+		[coder encodeValueOfObjCType:@encode(CGFloat) at:&(currentElement->position)];
 		
 		count++;
 		currentElement = currentElement->nextElement;
@@ -115,11 +115,11 @@ static void resolveHSV(float *color1, float *color2);
 	{
     CTGradientElement newElement;
 	
-	[coder decodeValueOfObjCType:@encode(float) at:&(newElement.red)];
-	[coder decodeValueOfObjCType:@encode(float) at:&(newElement.green)];
-	[coder decodeValueOfObjCType:@encode(float) at:&(newElement.blue)];
-	[coder decodeValueOfObjCType:@encode(float) at:&(newElement.alpha)];
-	[coder decodeValueOfObjCType:@encode(float) at:&(newElement.position)];
+	[coder decodeValueOfObjCType:@encode(CGFloat) at:&(newElement.red)];
+	[coder decodeValueOfObjCType:@encode(CGFloat) at:&(newElement.green)];
+	[coder decodeValueOfObjCType:@encode(CGFloat) at:&(newElement.blue)];
+	[coder decodeValueOfObjCType:@encode(CGFloat) at:&(newElement.alpha)];
+	[coder decodeValueOfObjCType:@encode(CGFloat) at:&(newElement.position)];
 	
 	count--;
 	[self addElement:&newElement];
@@ -418,7 +418,7 @@ static void resolveHSV(float *color1, float *color2);
   {
   id newInstance = [[[self class] alloc] init];
   
-  struct {float hue; float position; float width;} colorBands[4];
+  struct {CGFloat hue; CGFloat position; CGFloat width;} colorBands[4];
   
   colorBands[0].hue = 22;
   colorBands[0].position = .145;
@@ -440,7 +440,7 @@ static void resolveHSV(float *color1, float *color2);
   /////////////////////////////
   for(i = 0; i < 4; i++)
 	{	
-	float color[4];
+	CGFloat color[4];
 	color[0] = colorBands[i].hue - 180*colorBands[i].width;
 	color[1] = 1;
 	color[2] = 0.001;
@@ -494,7 +494,7 @@ static void resolveHSV(float *color1, float *color2);
 
 
 #pragma mark Modification
-- (CTGradient *)gradientWithAlphaComponent:(float)alpha
+- (CTGradient *)gradientWithAlphaComponent:(CGFloat)alpha
   {
   id newInstance = [[[self class] alloc] init];
   
@@ -525,7 +525,7 @@ static void resolveHSV(float *color1, float *color2);
 
 //Adds a color stop with <color> at <position> in elementList
 //(if two elements are at the same position then added imediatly after the one that was there already)
-- (CTGradient *)addColorStop:(NSColor *)color atPosition:(float)position
+- (CTGradient *)addColorStop:(NSColor *)color atPosition:(CGFloat)position
   {
   CTGradient *newGradient = [self copy];
   CTGradientElement newGradientElement;
@@ -545,7 +545,7 @@ static void resolveHSV(float *color1, float *color2);
 
 
 //Removes the color stop at <position> from elementList
-- (CTGradient *)removeColorStopAtPosition:(float)position
+- (CTGradient *)removeColorStopAtPosition:(CGFloat)position
   {
   CTGradient *newGradient = [self copy];
   CTGradientElement removedElement = [newGradient removeElementAtPosition:position];
@@ -592,9 +592,9 @@ static void resolveHSV(float *color1, float *color2);
   return nil;
   }
 
-- (NSColor *)colorAtPosition:(float)position
+- (NSColor *)colorAtPosition:(CGFloat)position
   {
-  float components[4];
+  CGFloat components[4];
   
   switch(blendingMode)
 	{
@@ -622,7 +622,7 @@ static void resolveHSV(float *color1, float *color2);
   [self fillRect:rect angle:45];
   }
 
-- (void)fillRect:(NSRect)rect angle:(float)angle
+- (void)fillRect:(NSRect)rect angle:(CGFloat)angle
   {
   //First Calculate where the beginning and ending points should be
   CGPoint startPoint;
@@ -640,14 +640,14 @@ static void resolveHSV(float *color1, float *color2);
   	}
   else						//ok, we'll do the calculations now 
   	{
-  	float x,y;
-  	float sina, cosa, tana;
+  	CGFloat x,y;
+  	CGFloat sina, cosa, tana;
   	
-  	float length;
-  	float deltax,
+  	CGFloat length;
+  	CGFloat deltax,
   		  deltay;
 	
-  	float rangle = angle * pi/180;	//convert the angle to radians
+  	CGFloat rangle = angle * pi/180;	//convert the angle to radians
 	
   	if(fabsf(tan(rangle))<=1)	//for range [-45,45], [135,225]
 		{
@@ -703,8 +703,8 @@ static void resolveHSV(float *color1, float *color2);
 - (void)radialFillRect:(NSRect)rect
   {
   CGPoint startPoint, endPoint;
-  float startRadius, endRadius;
-  float scalex, scaley, transx, transy;
+  CGFloat startRadius, endRadius;
+  CGFloat scalex, scaley, transx, transy;
   
   startPoint = endPoint = CGPointMake(NSMidX(rect), NSMidY(rect));
   
@@ -746,7 +746,7 @@ static void resolveHSV(float *color1, float *color2);
   CGContextRestoreGState(currentContext);
   }
 
-- (void)fillBezierPath:(NSBezierPath *)path angle:(float)angle
+- (void)fillBezierPath:(NSBezierPath *)path angle:(CGFloat)angle
   {
   NSGraphicsContext *currentContext = [NSGraphicsContext currentContext];
   [currentContext saveGraphicsState];
@@ -798,8 +798,8 @@ static void resolveHSV(float *color1, float *color2);
     
   CGFunctionCallbacks evaluationCallbackInfo = {0 , evaluationFunction, NULL};	//Version, evaluator function, cleanup function
   
-  static const float input_value_range   [2] = { 0, 1 };						//range  for the evaluator input
-  static const float output_value_ranges [8] = { 0, 1, 0, 1, 0, 1, 0, 1 };		//ranges for the evaluator output (4 returned values)
+  static const CGFloat input_value_range   [2] = { 0, 1 };						//range  for the evaluator input
+  static const CGFloat output_value_ranges [8] = { 0, 1, 0, 1, 0, 1, 0, 1 };		//ranges for the evaluator output (4 returned values)
   
   gradientFunction = CGFunctionCreate(&elementList,					//the two transition colors
 									  1, input_value_range  ,		//number of inputs (just fraction of progression)
@@ -880,7 +880,7 @@ static void resolveHSV(float *color1, float *color2);
   return removedElement;
   }
 
-- (CTGradientElement)removeElementAtPosition:(float)position
+- (CTGradientElement)removeElementAtPosition:(CGFloat)position
   {
   CTGradientElement removedElement;
   
@@ -949,9 +949,9 @@ static void resolveHSV(float *color1, float *color2);
 
 #pragma mark Core Graphics
 //////////////////////////////////////Blending Functions/////////////////////////////////////
-void linearEvaluation (void *info, const float *in, float *out)
+void linearEvaluation (void *info, const CGFloat *in, CGFloat *out)
   {
-  float position = *in;
+  CGFloat position = *in;
   
   if(*(CTGradientElement **)info == nil)	//if elementList is empty return clear color
 	{
@@ -1027,9 +1027,9 @@ void linearEvaluation (void *info, const float *in, float *out)
 //	this we will add to the hue's angle (if we subtract we'll be doing the inverse
 //	chromatic...scroll down more for that). All we need to do is keep adding to the hue
 //  until we wrap around the colorwheel and get to color2.
-void chromaticEvaluation(void *info, const float *in, float *out)
+void chromaticEvaluation(void *info, const CGFloat *in, CGFloat *out)
   {
-  float position = *in;
+  CGFloat position = *in;
   
   if(*(CTGradientElement **)info == nil)	//if elementList is empty return clear color
 	{
@@ -1041,8 +1041,8 @@ void chromaticEvaluation(void *info, const float *in, float *out)
   CTGradientElement *color1 = *(CTGradientElement **)info;
   CTGradientElement *color2 = color1->nextElement;
   
-  float c1[4];
-  float c2[4];
+  CGFloat c1[4];
+  CGFloat c2[4];
     
   //make sure first color and second color are on other sides of position
   while(color2 != nil && color2->position < position)
@@ -1115,9 +1115,9 @@ void chromaticEvaluation(void *info, const float *in, float *out)
 //	is strictly decreasing, that is we need to get from color1 to color2 by decreasing
 //	the 'angle' (i.e. 90¼ -> 180¼ would be done by subtracting 270¼ and getting -180¼...
 //	which is equivalent to 180¼ mod 360¼
-void inverseChromaticEvaluation(void *info, const float *in, float *out)
+void inverseChromaticEvaluation(void *info, const CGFloat *in, CGFloat *out)
   {
-    float position = *in;
+    CGFloat position = *in;
   
   if(*(CTGradientElement **)info == nil)	//if elementList is empty return clear color
 	{
@@ -1129,8 +1129,8 @@ void inverseChromaticEvaluation(void *info, const float *in, float *out)
   CTGradientElement *color1 = *(CTGradientElement **)info;
   CTGradientElement *color2 = color1->nextElement;
   
-  float c1[4];
-  float c2[4];
+  CGFloat c1[4];
+  CGFloat c2[4];
       
   //make sure first color and second color are on other sides of position
   while(color2 != nil && color2->position < position)
@@ -1196,15 +1196,15 @@ void inverseChromaticEvaluation(void *info, const float *in, float *out)
   }
 
 
-void transformRGB_HSV(float *components) //H,S,B -> R,G,B
+void transformRGB_HSV(CGFloat *components) //H,S,B -> R,G,B
 	{
-	float H=0, S=0, V=0;
+	CGFloat H=0, S=0, V=0;
 
-	float R = components[0],
+	CGFloat R = components[0],
 		  G = components[1],
 		  B = components[2];
 	
-	float MAX = R > G ? (R > B ? R : B) : (G > B ? G : B),
+	CGFloat MAX = R > G ? (R > B ? R : B) : (G > B ? G : B),
 	      MIN = R < G ? (R < B ? R : B) : (G < B ? G : B);
 	
 	if(MAX == MIN)
@@ -1227,15 +1227,15 @@ void transformRGB_HSV(float *components) //H,S,B -> R,G,B
 	components[2] = V;
 	}
 
-void transformHSV_RGB(float *components) //H,S,B -> R,G,B
+void transformHSV_RGB(CGFloat *components) //H,S,B -> R,G,B
 	{
-	float R=0, G=0, B=0;
-	float H = fmodf(components[0],359),	//map to [0,360)
+	CGFloat R=0, G=0, B=0;
+	CGFloat H = fmodf(components[0],359),	//map to [0,360)
 		  S = components[1],
 		  V = components[2];
 	
 	int   Hi = (int)floorf(H/60.) % 6;
-	float f  = H/60-Hi,
+	CGFloat f  = H/60-Hi,
 		  p  = V*(1-S),
 		  q  = V*(1-f*S),
 		  t  = V*(1-(1-f)*S);
@@ -1255,7 +1255,7 @@ void transformHSV_RGB(float *components) //H,S,B -> R,G,B
 	components[2] = B;
 	}
 
-void resolveHSV(float *color1, float *color2)	//H value may be undefined (i.e. graycale color)
+void resolveHSV(CGFloat *color1, CGFloat *color2)	//H value may be undefined (i.e. graycale color)
 	{											//	we want to fill it with a sensible value
 	if(isnan(color1[0]) && isnan(color2[0]))
 		color1[0] = color2[0] = 0;
