+++ Other Sources/Foreign Code/UKNiftyFeatures/UKPrefsPanel.m
@@ -82,7 +82,7 @@
 -(void)	awakeFromNib
 {
 	NSString*		key;
-	int				index = 0;
+	NSInteger		index = 0;
 	NSString*		wndTitle = nil;
 	
 	// Generate a string containing the window's title so we can display the original window title plus the selected pane:
@@ -130,7 +130,7 @@
 {
     // Create a new toolbar instance, and attach it to our document window 
     NSToolbar		*toolbar =[[tabView window] toolbar];
-	int				itemCount = 0,
+	NSInteger		itemCount = 0,
 					x = 0;
 	NSTabViewItem	*currPage = nil;
 	
@@ -316,7 +316,7 @@
 
 -(NSArray*) toolbarDefaultItemIdentifiers: (NSToolbar *) toolbar
 {
-	int					itemCount = [tabView numberOfTabViewItems],
+	NSInteger			itemCount = [tabView numberOfTabViewItems],
 						x;
 	NSTabViewItem*		theItem = [tabView tabViewItemAtIndex:0];
 	NSMutableArray*	defaultItems = [NSMutableArray array];
